requirements.txt
fyers-apiv3>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
python-dotenv>=1.0.0
joblib>=1.3.0
ta>=0.10.2
